Formatted header information
I've noticed that the presence of NAME_X or NAME_XX is inconsistent in
the headers. I've cut those and now the taxonomy string is either 5,6 or
7 names long. I've made it so the cut names are replaced with a blank
space so  the semicolons are preserved.

i.e.
>0001_Phatr11014-4Phatr2 TAXID="2850" ANNOTATION="Predicted protein
Fragment" SPECIES="Phaeodactylum tricornutum"
TAXONOMY="Eukaryota;Stramenopiles;Stramenopiles_X;Bacillariophyta;Bacillariophyta_X;Raphid-pennate;Phaeodactylum;Phaeodactylum
tricornutum"

becomes

>0001_"Eukaryota;Stramenopiles;;Bacillariophyta;;Raphid-pennate;Phaeodactylum;Phaeodactylum
tricornutum"

Also there is an inconsistent pattern where the first word of the
species begins with the genus so we had ;genus;genus species.  I made it
so if the genus matched the first word in the "genus species" exactly it
was replaced with a black space, and if it didn't match it was
preserved.  So the above becomes:

>0001_"Eukaryota;Stramenopiles;;Bacillariophyta;;Raphid-pennate;;Phaeodactylum
tricornutum"

But for instance:
>4158_gi119471972-NZ_AAVS01000028 TAXID="156578" ANNOTATION="Acyl-CoA
dehydrogenase, C-terminal domain" SPECIES="Alteromonadales bacterium
TW-7"
TAXONOMY="Bacteria;Proteobacteria;Gammaproteobacteria;Alteromonadales;Alteromonadaceae;Alteromonas;Alteromonadales
bacterium TW-7"

Becomes:

>4158_"Bacteria;Proteobacteria;Gammaproteobacteria;Alteromonadales;Alteromonadaceae;Alteromonas;Alteromonadales
bacterium TW-7"

For the step Search for identical amino acid sequences and then only
keep the longest sequence: Sorry, outlook sent the email while I was
typing it.

OK I'm on the step  Search for identical amino acid sequences and then
only keep the longest sequence:  and there are identical sequences.  If
we remove duplicates we go from 4211 sequences to 4137.  The keep
longest sequence instruction doesn't work because the sequences are
identical. I've kept the version of the sequence with the smallest ID
number and moved all duplicates to a second file.

OK:

REVISED_MASTERFILE.fa is the masterfile with our IDS prepended to the
header lines.
Trimmed_Headers.fa is the REVISED_MASTERFILE.fa with the redundant
information removed from the headers as described in the todo document.
Only_Unique_Aminos.fa is the set of fasta strings with only unique amino
acids and Redundant_Aminos.fa are all the fasta strings that were cut as
not unique.

I'm working on the reblasting the blast results and blast research steps
now.